product_recommender/requirements.txt
langgraph==0.0.34
youtube-search-python==1.6.6
yt-dlp==2024.4.9
openai-whisper==20231117
moviepy==1.0.3
textblob==0.17.1
streamlit==1.33.0
langchain-groq==0.1.2
httpx==0.26.0